Add separate MSW well pipe centerline computations
