+++ ApplicationLibCode/Application/RiaPreferences.h
@@ -86,7 +86,6 @@ public:
     bool        openExportedPdfInViewer() const;
     bool        useQtChartsAsDefaultPlotType() const;
     bool        writeEchoInGrdeclFiles() const;
-    bool        showSimplifiedMswWellPathGeometry() const;
 
     RiaDefines::ThemeEnum guiTheme() const;
 
@@ -218,7 +217,6 @@ private:
     caf::PdmField<bool>                                         m_enableFaultsByDefault;
     caf::PdmField<bool>                                         m_showInfoBox;
     caf::PdmField<bool>                                         m_showGridBox;
-    caf::PdmField<bool>                                         m_showSimplifiedMswWellPathGeometry;
 
     QStringList m_tabNames;
 };
