+++ ApplicationLibCode/ReservoirDataModel/CMakeLists_files.cmake
@@ -96,6 +96,7 @@ set(SOURCE_GROUP_HEADER_FILES
     ${CMAKE_CURRENT_LIST_DIR}/RigSurfaceStatisticsCalculator.h
     ${CMAKE_CURRENT_LIST_DIR}/RigWellLogIndexDepthOffset.h
     ${CMAKE_CURRENT_LIST_DIR}/RigPressureDepthData.h
+    ${CMAKE_CURRENT_LIST_DIR}/RigMswCenterLineCalculator.h
 )
 
 set(SOURCE_GROUP_SOURCE_FILES
@@ -189,6 +190,7 @@ set(SOURCE_GROUP_SOURCE_FILES
     ${CMAKE_CURRENT_LIST_DIR}/RigSurfaceStatisticsCalculator.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RigWellLogIndexDepthOffset.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RigPressureDepthData.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/RigMswCenterLineCalculator.cpp
 )
 
 list(APPEND CODE_HEADER_FILES ${SOURCE_GROUP_HEADER_FILES})
