+++ ApplicationLibCode/ReservoirDataModel/RigSimulationWellCenterLineCalculator.cpp
@@ -43,230 +43,6 @@
 //--------------------------------------------------------------------------------------------------
 ///
 //--------------------------------------------------------------------------------------------------
-std::vector<SimulationWellCellBranch> RigSimulationWellCenterLineCalculator::calculateMswWellPipeGeometry( RimSimWellInView* rimWell )
-{
-    CVF_ASSERT( rimWell );
-
-    const RigSimWellData* simWellData = rimWell->simWellData();
-    if ( !simWellData ) return {};
-
-    RimEclipseView* eclipseView;
-    rimWell->firstAncestorOrThisOfType( eclipseView );
-
-    CVF_ASSERT( eclipseView );
-
-    RigEclipseCaseData* eclipseCaseData = eclipseView->eclipseCase()->eclipseCaseData();
-
-    int timeStepIndex = eclipseView->currentTimeStep();
-
-    return calculateMswWellPipeGeometryForTimeStep( eclipseCaseData, simWellData, timeStepIndex );
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-std::vector<SimulationWellCellBranch>
-    RigSimulationWellCenterLineCalculator::calculateMswWellPipeGeometryForTimeStep( const RigEclipseCaseData* eclipseCaseData,
-                                                                                    const RigSimWellData*     wellResults,
-                                                                                    int                       timeStepIndex )
-{
-    if ( timeStepIndex >= 0 && !wellResults->hasAnyValidCells( timeStepIndex ) ) return {};
-
-    const RigWellResultFrame* wellFramePtr = nullptr;
-
-    if ( timeStepIndex < 0 )
-    {
-        wellFramePtr = wellResults->staticWellCells();
-    }
-    else
-    {
-        wellFramePtr = wellResults->wellResultFrame( timeStepIndex );
-    }
-
-    const RigWellResultFrame&               wellFrame      = *wellFramePtr;
-    const std::vector<RigWellResultBranch>& resultBranches = wellFrame.m_wellResultBranches;
-
-    std::vector<WellBranch> wellBranches = buildAccumulatedWellBranches( resultBranches );
-
-    // Connect outlet segment of branches to parent branch
-
-    for ( const auto& resultBranch : resultBranches )
-    {
-        if ( resultBranch.m_branchResultPoints.empty() ) continue;
-
-        const auto firstResultPoint = resultBranch.m_branchResultPoints.front();
-
-        for ( auto& wellBranch : wellBranches )
-        {
-            if ( wellBranch.m_branchId == resultBranch.m_ertBranchId )
-            {
-                if ( firstResultPoint.branchId() == resultBranch.m_ertBranchId )
-                {
-                    // The first result point is on the same branch, use well head as outlet
-                    RigWellResultPoint outletResultPoint = wellFrame.m_wellHead;
-
-                    auto gridAndCellIndex = std::make_pair( outletResultPoint.gridIndex(), outletResultPoint.cellIndex() );
-                    wellBranch.m_segmentsWithGridCells[outletResultPoint.segmentId()].push_back( gridAndCellIndex );
-                }
-                else
-                {
-                    // The first result point on a different branch. Find the branch and add the grid cell
-                    for ( const auto& candidateResultBranch : wellBranches )
-                    {
-                        if ( firstResultPoint.branchId() == candidateResultBranch.m_branchId )
-                        {
-                            std::pair<size_t, size_t> gridAndCellIndexForTieIn;
-
-                            for ( const auto& [segment, gridAndCellIndices] : candidateResultBranch.m_segmentsWithGridCells )
-                            {
-                                if ( segment > firstResultPoint.segmentId() ) continue;
-                                if ( !gridAndCellIndices.empty() )
-                                {
-                                    gridAndCellIndexForTieIn = gridAndCellIndices.front();
-                                }
-                            }
-
-                            wellBranch.m_segmentsWithGridCells[firstResultPoint.segmentId()].push_back( gridAndCellIndexForTieIn );
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    std::vector<SimulationWellCellBranch> simWellBranches;
-    for ( const auto& wellBranch : wellBranches )
-    {
-        std::vector<cvf::Vec3d>         branchCoords;
-        std::vector<RigWellResultPoint> resultPoints;
-
-        if ( wellBranch.m_branchId == 1 && !wellBranch.m_segmentsWithGridCells.empty() )
-        {
-            const auto& [firstSegment, gridAndCellIndices] = *wellBranch.m_segmentsWithGridCells.begin();
-            if ( !gridAndCellIndices.empty() )
-            {
-                const auto& [gridIndex, cellIndex] = gridAndCellIndices.front();
-
-                const RigCell& cell       = eclipseCaseData->grid( gridIndex )->cell( cellIndex );
-                cvf::Vec3d     whStartPos = cell.faceCenter( cvf::StructGridInterface::NEG_K );
-
-                // Add extra coordinate between cell face and cell center
-                // to make sure the well pipe terminated in a segment parallel to z-axis
-
-                cvf::Vec3d whIntermediate = whStartPos;
-                whIntermediate.z()        = ( whStartPos.z() + cell.center().z() ) / 2.0;
-
-                RigWellResultPoint resPoint;
-                for ( const auto& resBranch : resultBranches )
-                {
-                    for ( const auto& respoint : resBranch.m_branchResultPoints )
-                    {
-                        if ( respoint.segmentId() == firstSegment )
-                        {
-                            resPoint = respoint;
-                            break;
-                        }
-                    }
-                }
-
-                branchCoords.push_back( whStartPos );
-                resultPoints.push_back( resPoint );
-                branchCoords.push_back( whIntermediate );
-                resultPoints.push_back( resPoint );
-            }
-        }
-
-        for ( const auto& [segmentId, gridAndCellIndices] : wellBranch.m_segmentsWithGridCells )
-        {
-            for ( const auto& [gridIndex, cellIndex] : gridAndCellIndices )
-            {
-                const RigCell& cell = eclipseCaseData->grid( gridIndex )->cell( cellIndex );
-                cvf::Vec3d     pos  = cell.center();
-
-                RigWellResultPoint resPoint;
-
-                // The result point is only used to transport the grid index and cell index
-                // The current implementation will propagate the cell open state to the well segment from one cell to
-                // the next. It this is misleading, the two following lines can be removed.
-                resPoint.setGridIndex( gridIndex );
-                resPoint.setGridCellIndex( cellIndex );
-
-                branchCoords.push_back( pos );
-                resultPoints.push_back( resPoint );
-            }
-        }
-
-        branchCoords.push_back( branchCoords.back() );
-
-        simWellBranches.emplace_back( branchCoords, resultPoints );
-    }
-
-    return simWellBranches;
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-std::vector<RigSimulationWellCenterLineCalculator::WellBranch>
-    RigSimulationWellCenterLineCalculator::buildAccumulatedWellBranches( const std::vector<RigWellResultBranch>& resBranches )
-{
-    std::vector<WellBranch> longWellBranches;
-    std::vector<WellBranch> shortWellBranches;
-
-    for ( const auto& resultBranch : resBranches )
-    {
-        WellBranch branch;
-        branch.m_branchId = resultBranch.m_ertBranchId;
-
-        for ( const auto& resPoint : resultBranch.m_branchResultPoints )
-        {
-            size_t gridIndex     = resPoint.gridIndex();
-            size_t gridCellIndex = resPoint.cellIndex();
-
-            auto gridAndCellIndex = std::make_pair( gridIndex, gridCellIndex );
-            if ( gridIndex != cvf::UNDEFINED_SIZE_T && gridCellIndex != cvf::UNDEFINED_SIZE_T && !branch.containsGridCell( gridAndCellIndex ) )
-            {
-                OutputSegment outputSegment{ resPoint.outletSegmentId(), resPoint.outletBranchId() };
-                branch.m_gridCellsConnectedToSegments[gridAndCellIndex] = outputSegment;
-                branch.m_segmentsWithGridCells[resPoint.segmentId()].push_back( gridAndCellIndex );
-            }
-        }
-
-        const int resultPointThreshold = 3;
-        if ( resultBranch.m_branchResultPoints.size() > resultPointThreshold )
-        {
-            longWellBranches.push_back( branch );
-        }
-        else
-        {
-            shortWellBranches.push_back( branch );
-        }
-    }
-
-    // Move all grid cells of small branch to the long branch
-    for ( const auto& branch : shortWellBranches )
-    {
-        if ( branch.m_gridCellsConnectedToSegments.empty() ) continue;
-
-        auto outputSegment = branch.m_gridCellsConnectedToSegments.begin()->second;
-        for ( auto& longBranch : longWellBranches )
-        {
-            if ( longBranch.m_branchId == outputSegment.outputSegmentBranchId )
-            {
-                for ( const auto& [gridAndCellIndex, localOutputSegment] : branch.m_gridCellsConnectedToSegments )
-                {
-                    longBranch.m_segmentsWithGridCells[outputSegment.outputSegmentId].push_back( gridAndCellIndex );
-                }
-            }
-        }
-    }
-
-    return longWellBranches;
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
 std::vector<SimulationWellCellBranch> RigSimulationWellCenterLineCalculator::calculateWellPipeStaticCenterline( RimSimWellInView* rimWell )
 {
     std::vector<std::vector<cvf::Vec3d>>         pipeBranchesCLCoords;
