+++ ApplicationLibCode/ReservoirDataModel/RigSimulationWellCenterLineCalculator.h
@@ -39,7 +39,6 @@ class RigWellResultFrame;
 class RigSimulationWellCenterLineCalculator
 {
 public:
-    static std::vector<SimulationWellCellBranch> calculateMswWellPipeGeometry( RimSimWellInView* rimWell );
     static std::vector<SimulationWellCellBranch> calculateWellPipeStaticCenterline( RimSimWellInView* rimWell );
 
     static std::vector<SimulationWellCellBranch> calculateWellPipeCenterlineForTimeStep( const RigEclipseCaseData* eclipseCaseData,
@@ -50,34 +49,6 @@ public:
 
     static std::pair<std::vector<std::vector<cvf::Vec3d>>, std::vector<std::vector<RigWellResultPoint>>>
         extractBranchData( const std::vector<SimulationWellCellBranch> simulationBranch );
-
-private:
-    struct OutputSegment
-    {
-        int outputSegmentId       = -1;
-        int outputSegmentBranchId = -1;
-    };
-
-    struct WellBranch
-    {
-        int m_branchId = -1;
-
-        std::map<std::pair<size_t, size_t>, OutputSegment>    m_gridCellsConnectedToSegments;
-        std::map<int, std::vector<std::pair<size_t, size_t>>> m_segmentsWithGridCells;
-
-        bool containsGridCell( const std::pair<size_t, size_t>& candidateGridAndCellIndex ) const
-        {
-            for ( const auto& [segmentId, gridAndCellIndex] : m_segmentsWithGridCells )
-            {
-                if ( std::find( gridAndCellIndex.begin(), gridAndCellIndex.end(), candidateGridAndCellIndex ) != gridAndCellIndex.end() )
-                {
-                    return true;
-                }
-            }
-
-            return false;
-        };
-    };
 
 private:
     static void calculateWellPipeStaticCenterline( RimSimWellInView*                             rimWell,
@@ -96,7 +67,9 @@ private:
                                                                                           const RigSimWellData*     simWellData,
                                                                                           int                       timeStepIndex );
 
-    static std::vector<WellBranch> buildAccumulatedWellBranches( const std::vector<RigWellResultBranch>& resBranches );
+    static SimulationWellCellBranch addSegmentsToCellFaces( const std::vector<cvf::Vec3d>          branchCoords,
+                                                            const std::vector<RigWellResultPoint>& resultPoints,
+                                                            const RigEclipseCaseData*              eclipseCaseData );
 
     static bool hasAnyValidDataCells( const RigWellResultBranch& branch );
     static void finishPipeCenterLine( std::vector<std::vector<cvf::Vec3d>>& pipeBranchesCLCoords, const cvf::Vec3d& lastCellCenter );
